requirements.txt
discord.py
requests
flask
google-api-python-client
